move use case configuration into application package
