--- api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/order/OrderConfiguration.kt
@@ -1,10 +1,0 @@
-package github.com.hukuta94.delivery.api.startup.configuration.order
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Import
-
-@Configuration
-@Import(
-    OrderUseCaseConfiguration::class,
-)
-open class OrderConfiguration
